+++ src/materials/Lambertian.cpp
@@ -9,20 +9,20 @@ namespace raytracer
 {
 //----------------------------------------------------------------------------------
 Lambertian::Lambertian()
-    : m_albedo(0.0f, 0.0f, 0.0f)
+    : m_albedo(0.0, 0.0, 0.0)
 {
 }
 
 //----------------------------------------------------------------------------------
-Lambertian::Lambertian(const glm::vec3 &albedo)
+Lambertian::Lambertian(const glm::dvec3 &albedo)
     : m_albedo(albedo)
 {
 }
 
 //----------------------------------------------------------------------------------
-bool Lambertian::scatter(const Ray &ray, const HitRecord &record, glm::vec3 &attenuation, Ray &scattered) const
+bool Lambertian::scatter(const Ray &ray, const HitRecord &record, glm::dvec3 &attenuation, Ray &scattered) const
 {
-    glm::vec3 scatterDirection = record.normal + glm::normalize(randomInUnitSphere());
+    glm::dvec3 scatterDirection = record.normal + glm::normalize(randomInUnitSphere());
 
     // Catch degenerate scatter direction
     if(nearZero(scatterDirection))
